+++ builtin/credential/app-id/backend_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	logicaltest "github.com/hashicorp/vault/sdk/acctest"
+	logicaltest "github.com/hashicorp/vault/sdk/testing/stepwise"
 	"github.com/hashicorp/vault/sdk/helper/salt"
 	"github.com/hashicorp/vault/sdk/logical"
 )
