+++ builtin/credential/github/backend_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	logicaltest "github.com/hashicorp/vault/sdk/acctest"
+	logicaltest "github.com/hashicorp/vault/sdk/testing/stepwise"
 	"github.com/hashicorp/vault/sdk/logical"
 )
