+++ builtin/credential/userpass/backend_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/go-test/deep"
 	sockaddr "github.com/hashicorp/go-sockaddr"
-	logicaltest "github.com/hashicorp/vault/sdk/acctest"
+	logicaltest "github.com/hashicorp/vault/sdk/testing/stepwise"
 	"github.com/hashicorp/vault/sdk/helper/policyutil"
 	"github.com/hashicorp/vault/sdk/helper/tokenutil"
 	"github.com/hashicorp/vault/sdk/logical"
