+++ builtin/logical/ssh/backend_test.go
@@ -17,7 +17,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/vault/api"
-	logicaltest "github.com/hashicorp/vault/sdk/acctest"
+	logicaltest "github.com/hashicorp/vault/sdk/testing/stepwise"
 	"github.com/hashicorp/vault/sdk/logical"
 	"github.com/hashicorp/vault/vault"
 	"github.com/mitchellh/mapstructure"
