+++ builtin/logical/transit/backend_test.go
@@ -16,7 +16,7 @@ import (
 	"time"
 
 	uuid "github.com/hashicorp/go-uuid"
-	logicaltest "github.com/hashicorp/vault/sdk/acctest"
+	logicaltest "github.com/hashicorp/vault/sdk/testing/stepwise"
 	"github.com/hashicorp/vault/sdk/framework"
 	"github.com/hashicorp/vault/sdk/helper/keysutil"
 	"github.com/hashicorp/vault/sdk/logical"
