+++ sdk/testing/stepwise/testing_test.go
@@ -1,4 +1,4 @@
-package acctest
+package stepwise
 
 import (
 	"os"
